:sparkles: Merge `feature-migration-waves` to `main` (disabled behind feature flag) (#711)
This PR is a placeholder to remind us that we have an ongoing feature
branch in `feature-migration-waves` (started in #598). Other PRs will be
merged to that branch, and then we can rebase and merge this PR when
we're ready to put all those changes into the main branch.

Changes included here are from:
* #598
* #776
* #740
* #778
* #790

See notes in https://github.com/konveyor/tackle2-ui/pull/740 for some
next steps.

Edit: we've added a feature flag in #790 so we can merge this to main
even though it's incomplete.

---------

Signed-off-by: Ian Bolton <ibolton@redhat.com>
Signed-off-by: Mike Turley <mike.turley@alum.cs.umass.edu>
Co-authored-by: Ian Bolton <ibolton@redhat.com>