:seedling: Replace render-props-based `ComposableTableWithControls` with hook-based `useTableControls` (#808)
Followup from #792. Refactoring that component to be more composable led
to a discussion on how we could lean even harder into modern react
composition-over-configuration principles, using React Hooks to drive
the entire abstraction.

Note: This PR also adds generics (a type param for the item objects
being selected) to the `ToolbarBulkSelector` component. This change
caused a type error in the AdoptionCandidateTable, where
`currentPageItems` was being passed an array of row configuration
objects instead of an array of applications. **This revealed a bug in
that table: Choosing "Select page" from that bulk selector didn't
work**. This PR fixes the bug by mapping over the objects to pass in an
array of applications, as expected by the component.

---------

Signed-off-by: Mike Turley <mike.turley@alum.cs.umass.edu>