:bug: Use a React key prop to remount the drawer contents when changing routes (#851)
Followup to #848.

This solves the weird behavior where navigating away from a page of the
app that has a drawer will cause the drawer to slide closed after
navigating rather than disappear immediately with the rest of the page.
This fixes an additional quirk I encountered while introducing a drawer
to the Dependencies page: I would open the drawer for a Dependency,
switch to the apps page, and right as the drawer started closing it
would swap its contents for the app table's drawer contents.

A `key` prop on any React component will cause the component to fully
unmount and remount when the key changes. This PR adds a `pageKey` prop
to the `PageDrawerContent` component, to be a string literal describing
the page where a drawer is used. This key is passed to the
`DrawerPanelContent` component, which is what has the CSS animation for
sliding the drawer open and closed. Remounting this component snaps the
drawer closed with no animation, but does not remount the entire page
contents (like we used to before #848).

Signed-off-by: Mike Turley <mike.turley@alum.cs.umass.edu>