:sparkles: Add support for clickable/active rows with URL param + example usage with Drawer on Dependencies table (#852)
Implements clickable rows and a drawer on the Dependencies table (with
placeholder content) via new reusable features in the table-control
hooks. The drawer on these new tables can now be trivially driven by a
URL param (thanks to `idProperty` being a thing) so when you navigate
away from a page with a drawer open and click Back, the open drawer is
restored :slightly_smiling_face:

* Adds `@app/shared/hooks/table-controls/utils.ts` for general
table-related utils that can be used either with or without the table
control hooks.
* Currently only contains `handlePropagatedRowClick`, which is factored
out from the clickable row / drawer changes in the applications table
introduced in https://github.com/konveyor/tackle2-ui/pull/654. This is
the handler that makes sure that when the user clicks interactable
elements within a row (buttons, checkboxes, etc) we don't trigger the
row click / drawer. It is now used both in the legacy tables and the new
tables.
* Adds an `active-row` folder to `@app/shared/hooks/table-controls`,
with:
* `useActiveRowState.ts` - similar to other table control concerns,
contains hooks for tracking the active row ID string in either React
state or a URL param
* `getActiveRowDerivedState.ts` - given the state and the
`currentPageItems` derived from existing hooks / query data, returns the
active row's actual `TItem` object (or null) for convenience. Also
returns simplified functions for setting/clearing the active row by item
rather than by id so the page level components don't have to do any id
comparisons.
* Note: this is different from the other `getLocal*DerivedState` hooks
(note the lack of "local" in the name), because it is used in
`useTableControlProps` no matter whether the table uses local or
server-side filtering/pagination/sorting logic. We store this state in a
URL param on the client, but it is not passed to the hub; deriving the
active row's item can always be done locally.
* `useActiveRowEffects.ts` - provides a `useEffect` that watches for a
corner case where the `activeRowId` in state no longer matches anything
present in `currentPageItems`, and clears the state.
* This avoids a weird quirk where you could click a row and open its
drawer, use pagination/filtering to make that row disappear (closing the
drawer), then go back to a view which includes the row and the drawer
would spontaneously reopen. Worse, if the row never reappeared at all
(e.g. it was deleted and the data was refetched), the stale
`activeRowId` would remain in state indefinitely.
* With this effect in place in `useTableControlProps`, once the drawer
closes for any reason it stays closed until the user clicks another row,
and the `activeRowId` in state always matches an item in
`currentPageItems`.
* There is an exception: If `isLoading` is true, we let the stale
`activeRowId` stick around until loading finishes so we can restore an
open drawer via URL (e.g. clicking the Back button) if the first data we
fetch includes a matching row.
* Adds `activeRowState` to the object returned by both
`useTableControlState` and `useTableControlUrlParams`, so it is
automatically consumed via the `args` object in `useTableControlProps`
for both styles of table.
* Adds a boolean `hasClickableRows` property to the
`ITableControlCommonArgs` so it can be used for whatever logic necessary
(currently only adds the `hasSelectableRowCaption` prop to
TableComposable in `useTableControlProps`).
* In `useTableControlProps`:
  * Adds calls to `getActiveRowDerivedState` and `useActiveRowEffects`
  * Adds a new propHelper `getClickableTrProps`
* This can be used either for clickable rows that set an active row, or
to make clickable rows for any reason (e.g. direct route navigation).
You can pass only an `item` to it to have it handle all the active row
state for you, or you can pass only an `onRowClick` to it for it to
handle and it won't touch the active row state. The benefit of this over
directly passing `onRowClick` to the `Tr` yourself is that you'll get
the `handlePropagatedRowClick` behavior.
* In the Dependencies table:
  * Adds `hasClickableRows: true`
  * Calls `getClickableTrProps` on each `Tr`
* Pulls `activeRowItem` and `clearActiveRow` out from
`activeRowDerivedState` and uses them to render the PageDrawerContent
* Drawer content is a placeholder for now and will be filled out in a
followup PR.
* Removes the `ConditionalRender` wrapping the table: after some
discussion on Slack with @gildub, it appears we don't really need this
anymore if the tables can handle loading states gracefully. We should
probably remove it elsewhere.

---------

Signed-off-by: Mike Turley <mike.turley@alum.cs.umass.edu>