:seedling: Gracefully shutdown the server on SIGINT/SIGTERM (#890)
Add graceful shutdown handling for Ctrl-C (SIGINT) and the podman/docker
container stop (SIGTERM) signals. There are a few advantages doing this:

  1. Allow connections to terminate, and

  2. Allow for any resource cleanup, and

3. Allow Ctrl-C (and `podman container stop` without resorting to
SIGKILL) to actually work when running the server as PID 1 in the
container.

For some graceful shutdown of node/express as PID 1 in a container
references, see:
  - https://expressjs.com/en/advanced/healthcheck-graceful-shutdown.html
-
https://github.com/nodejs/docker-node/blob/main/docs/BestPractices.md#handling-kernel-signals
-
https://emmer.dev/blog/you-don-t-need-an-init-system-for-node.js-in-docker/#express-doesn-t-trap-exit-signals

<!--
## PR Title Prefix

Every **PR Title** should be prefixed with :text: to indicate its type.

- Breaking change: :warning: (`:warning:`)
- Non-breaking feature: :sparkles: (`:sparkles:`)
- Patch fix: :bug: (`:bug:`)
- Docs: :book: (`:book:`)
- Infra/Tests/Other: :seedling: (`:seedling:`)
- No release note: :ghost: (`:ghost:`)

For example, a pull request containing breaking changes might look like
`:warning: My pull request contains breaking changes`.

Since GitHub supports emoji aliases (ie. `:ghost:`), there is no need to
include
the emoji directly in the PR title -- **please use the alias**. It used
to be
the case that projects using emojis for PR typing had to include the
emoji
directly because GitHub didn't render the alias. Given that `:warning:`
is
easy enough to read as text, easy to parse in release tooling, and
rendered in
GitHub well, we prefer to standardize on the alias.

For more information, please see the Konveyor
[Versioning
Doc](https://github.com/konveyor/release-tools/blob/main/VERSIONING.md).
-->

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>