:seedling: Add `npm run test` and update github action `ci-actions.yml` to use it
Add a `npm run test` script to the root package and adjust the
Github Action to run from the root directory.  This allows for
proper installation of any shared dependencies in the root `package.json`,
and for test scripts to be run in each workspace.

Note: The cli options given in ci-actions.yml for the test script
were not actually being passed to the jest command.  An extra `--`
is needed to send those options.  Those adjustments have also been
made.

Note 2: Using `npm clean-install` in the ci-actions.yml will cause
ci tests to fail if `package-lock.json` has not been updated in a PR.
This version of install prevents npm from updating the lock file.

Note 3: `.gitignore` updated for non-root test, build and dist paths

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>