:seedling: Refactor npm scripts, prefer coordination from root package
Change highlights:

  - Script `npm run build:dev` is dropped since it doesn't work. The
    current `_env` injection configurations in `webpack.dev.ts` is not
    compatible with serving the app from the express server.

  - Script `npm run start` builds the production client and starts the
    server, mimicking how the UI runs in container

  - Script `npm run start:dev` is managed at the root `package.json`

  - Script `npm run port-forward` is managed at the root `package.json`

The scripts moved to the root `package.json` allows for the reduction
of the scope of concern for each packages' scripts.  Now, the client only
knows how to run itself.  The root knows how to coordinate the various
pieces (the port mapping, the client's webpack-dev-server, and the
server's server) to get everything running.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>