:sparkles: Issues and Affected Apps: compatibility with new hub changes, carry app name filter between pages, use issues data as basis for table control hooks (#893)
This gets us a little closer to a complete issues / affected apps flow.
It's not complete, but I reached a decent stopping point.

Notes in order of appearance in the diff:

* Updates text and ids to match Issues and Affected Applications page
titles correctly
* Updates route for Affected Apps page to `/issues/:ruleset/:rule` (the
composite issue object no longer has a unique identifier (previously
ruleID), it is now identified by a unique pair of ruleset and rule)
* Renames the type of `AnalysisCompositeIssue` as returned from the hub
to `BaseAnalysisCompositeIssue`, so that we can have a type
`AnalysisCompositeIssue` which extends that with a new property
`_ui_unique_id`. This property is injected into these objects in the
`select` function of `useFetchCompositeIssues`. This is because our
table hooks expect an `idProperty` which can uniquely identify a row for
the purposes of state management. `_ui_unique_id` is just a combination
of ruleset and rule.
* Comments out code related to displaying source and target
technologies, and leaves TODO comments for restoring them. Technologies
have been moved into special labels in the hub, and we need to parse
them in a future change. That seemed out of scope for this PR.
* Removes sorting the affected apps page by application name, as this is
causing errors in the hub. TODO comments added to see if we can fix
this, need to check with @jortel.
* Adds a filter on application name to the affected apps page. This
filter is special: when moving from the Issues page to Affected Apps, if
there is an app name filter on the issues page we include it in the URL
and carry it over to the affected apps page. We also include a property
`fromIssuesParams` which is an encoded string of all the URL params that
were present on the issues page before we navigated away. Then when
getting the URL for the breadcrumb bar that returns us to the issues
page, we restore those params in `fromIssuesParams`, and if the user has
modified or cleared the app name filter while on the affected apps page,
we carry that change back over to the issues page. This filter-carrying
behavior does not conflict with any other filters or params which may be
present on either of these pages; everything will be retained and merged
together.
* This logic is defined in the `getAffectedAppsUrl` and
`getBackToIssuesUrl` functions in
`client/src/app/pages/issues/helpers.ts`. Note that the constant
`filterKeysToCarry` at the top can be modified to include more filters
in this carry-over behavior if we want to add more filters in the future
that are present on both of these pages.
* Adds an `implicitFilters` argument to `getHubRequestParams`. This can
be used to inject additional filters in the hub request that are not
visible to the user. On the affected apps page, we use it to filter the
hub's issues endpoint by the current ruleset and rule. Implicit filters
and the user-exposed filters get merged together when making a hub
request.
* Refactors the affected apps page so that all table control state is
based on issue objects rather than application objects (removes the use
of `useLocalTableControls`). Instead, the application ref on each issue
is resolved in the render logic inside the table JSX. This way we don't
need to worry about multiple sets of state for
sorting/filtering/pagination etc, because we need to apply those things
to the hub for issues and not apps anyway.
* Adds a TODO to return to about making `expansionState` and
`selectionState` optional in `useTableControlProps`. We have to include
a stub `selectionState` here even though it isn't being used. Fixing
this felt out of scope for this PR.
* Passes the composite issue's name in an additional URL param down to
the affected apps page. This is purely for rendering in the breadcrumb
bar so we have a consistent information architecture (the user should
see the name of the thing they drilled down from).
* Also includes the ruleset and rule in the breadcrumb bar for now,
since those are what make the view unique. This is not in the mockups:
we should discuss this with Justin in a future design review.
* On the issues page:
* Replaces the old Rule ID column with Name, and adds Rule set and Rule
columns for now (so the rows are unique). We should also discuss this in
a design review.
* Switches everything that depended on `idProperty: "ruleID"` to use
`_ui_unique_id`.
* Renames `issue` to `compositeIssue` in the table rendering code for
clarity.
* As mentioned above, adds a `select` option to `useQuery` in
`useFetchCompositeIssues` in order to inject `_ui_unique_id` on the
composite issue objects.
* For the `idProperty` field's type everywhere it is used
(`KeyWithValueType`), allows either a string or a number value for that
property. This change was made for accommodating the ruleID in certain
places before that was removed, but I left it because it's a useful
change.
* As mentioned above, adds the `implicitFilters` support to
`getFilterHubRequestParams`.
* In `useFilterState.ts`, factors out and exports the `serialize` and
`deserialize` functions from `useFilterUrlParams` so they can be used on
their own in the filter-carrying behavior of `getAffectedAppsUrl` and
`getBackToIssuesUrl` mentioned above. The new functions
`serializeFilterUrlParams` and `deserializeFilterUrlParams` can now be
used to manipulate filters in a URL string anywhere and retain the
special behavior (it is mostly just a JSON serialization, but
null/undefined/empty filters get automatically stripped out).
* In `useSortState`, makes `sortableColumns` optional because we may not
want sorting in every table. Currently there are no sortable columns in
the affected apps table due to the weirdness with querying issues but
displaying applications.
* In `useCompoundExpansionState`, casts the `item[idProperty]` as a
string wherever it is used in state in order to support the earlier
change that allows numeric id properties without modifying a bunch of
types that expect string keys in state.
* In `useUrlParams.ts`, factors out and exports a function
`trimAndStringifyUrlParams` so the behavior of excluding params from the
URL if they are null or undefined can be reused in `getAffectedAppsUrl`
and `getBackToIssuesUrl` above.

---------

Signed-off-by: Mike Turley <mike.turley@alum.cs.umass.edu>