:sparkles: Add support for additional useSelectionState args in useLocalTableControls (#901)
Currently `useLocalTableControlState` (and by extension
`useLocalTableControls`) uses `useSelectionState` internally and only
passes the required arguments to it. This PR exposes the optional
`initialSelected` and `isItemSelectable` arguments for
`useSelectionState` on the table args object.

Note: This is to unblock @gildub for now, but eventually we probably
want to refactor these to get rid of the usage of `useSelectionState` in
its current form, so that we can drive the state from string keys via
`idProperty` instead of references to row data in memory. This will make
it possible to lift selection state into the URL bar and to set items as
initially selected before the data is loaded.

Signed-off-by: Mike Turley <mike.turley@alum.cs.umass.edu>