:bug: Prevent infinite render loop when `drawerPanelContentProps` is omitted in PageDrawerContent (#986)
Fixes https://github.com/konveyor/tackle2-ui/issues/984

In `PageDrawerContent`, there are a few props the consumer can pass in
locally to the component that we sync up to the Context using
useEffects. The existing ones `isExpanded` and `pageKey` are primitives
(boolean and string) and so they were safe to use as dependencies in
`useEffect` since an unchanged value is referentially equal to the
previous value (`"foo" === "foo"`, `true === true`). However, in
https://github.com/konveyor/tackle2-ui/pull/904 I introduced
`drawerPanelContentProps`, which is an object with a default value of
`{}` when omitted. Because `{} !== {}`, when this prop was omitted the
useEffect to sync it to context was causing an infinite loop (React
thinks every render has a new value for it and updates context with that
new `{}` value, which triggers another render, etc).

The solution is not to default it to an object literal in the component
function's definition, but instead let it be undefined if the consumer
omits it and use `{}` as a fallback value when setting it in context
(will happen only once if the prop is omitted).