Credential modal, reset value and revalidate form for password onFocus
When `onFocus` to a password or token input field, reset the value
back to default and have the form revalidate. Use `resetField()`
instead of `setValue("password", "")` to achieve this behavior.

The field reset and form invalidation causes the Create button to be
disabled and the old password removed when focus enters the password
field.

Bug-Url: https://issues.redhat.com/browse/MTA-474
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>