:bug: Fix broken axios get calls for fetching tags and tag categories by id (#1005)
In https://github.com/konveyor/tackle2-ui/pull/889, the `getTagById` and
`getTagCategoryById` functions were changed to use axios, but the
`.then` callback was left out. That means a Promise is being returned
instead of a Tag or TagCategory, and TypeScript isn't catching it
because axios doesn't know what type it is fetching. This led to tags
not being displayed in the app details drawer on the app inventory page.