:bug: Update jira instance connection column to use popover (#1049)
When a jira instance on the main jira configuration table has a
connectivity issue, the connection column will show "Not connected".
Clicking the text will now open a popover to display additional details
about the state of the instance. The popover replaces a standard modal
box with similar information.

Fixes: #855

What it looked like before:

![image](https://github.com/konveyor/tackle2-ui/assets/3985964/1636600f-799c-491e-b3b7-6c0db7724edf)

What it looks like now:

![image](https://github.com/konveyor/tackle2-ui/assets/3985964/ce5422fb-0134-4a6f-a4b8-50969d7598c7)